Wrote method to query the db by letter
Wrote a method to query the database and return an array of english phrases beginning with the case-insensitive letter parameter passed into the method. The idea being that it will connect to the menu on the frontend with all the letters.
Not really sure how to query a mongoDB database so I'd appreciate it if someone with a bit more knowledge had a look over it.